+++ libs/framework-core/release_notes.md
@@ -3,7 +3,12 @@
 
 ## upcoming
 
-* [new] Client, Host; Run in main thread decorator used to run methods in the main thread.
+* [new] Client, Host; Using delegate_to_main_thread_wrapper decorator to execute methods in main thread function provided as run_in_main_thread_wrapper argument when instantiating.
+
+
+## v3.0.0rc1
+2024-07-17
+
 * [feature] Client, Engine, Dialog; Support tool config top level options.
 * [fix] Engine; Check enabled/disabled plugins.
 * [change] Client, Dialog; Support set_tool_config_option for any item in the tool_conifg.